snapshot of work in progress on separating out terminal tab from display of zero or more terminals
