+++ src/gwt/src/org/rstudio/studio/client/workbench/views/terminal/TerminalTab.java
@@ -17,39 +17,44 @@ package org.rstudio.studio.client.workbench.views.terminal;
 
 import com.google.inject.Inject;
 
-import org.rstudio.core.client.widget.model.ProvidesBusy;
+import org.rstudio.core.client.command.CommandBinder;
+import org.rstudio.core.client.command.Handler;
 import org.rstudio.studio.client.application.events.EventBus;
-import org.rstudio.studio.client.workbench.events.BusyHandler;
+import org.rstudio.studio.client.workbench.commands.Commands;
 import org.rstudio.studio.client.workbench.model.Session;
 import org.rstudio.studio.client.workbench.ui.DelayLoadTabShim;
 import org.rstudio.studio.client.workbench.ui.DelayLoadWorkbenchTab;
 import org.rstudio.studio.client.workbench.views.terminal.events.CreateTerminalEvent;
 
 public class TerminalTab extends DelayLoadWorkbenchTab<TerminalTabPresenter> 
-                         implements ProvidesBusy
 {
-   public abstract static class Shim extends
-                DelayLoadTabShim<TerminalTabPresenter, TerminalTab>
-      implements ProvidesBusy,
-                 CreateTerminalEvent.Handler
+   public interface Binder extends CommandBinder<Commands, Shim> {}
+
+   public abstract static class Shim 
+      extends DelayLoadTabShim<TerminalTabPresenter, TerminalTab>
+      implements CreateTerminalEvent.Handler
    {
+      @Handler
+      public abstract void onActivateTerminal();
+      
       abstract void initialize();
    }
 
    @Inject
-   public TerminalTab(Shim shim, EventBus events, final Session session)
+   public TerminalTab(Shim shim,
+                      EventBus events,
+                      Commands commands,
+                      Binder binder,
+                      final Session session)
    {
       super("Terminal", shim);
       shim_ = shim;
-
-      events.addHandler(CreateTerminalEvent.TYPE, shim);
+      
+      binder.bind(commands, shim_);
+      events.addHandler(CreateTerminalEvent.TYPE, shim_);
+      
+      shim_.initialize();
    }
 
-   @Override
-   public void addBusyHandler(BusyHandler handler)
-   {
-      shim_.addBusyHandler(handler);
-   }
-   
    private Shim shim_;
 }
